+++ CHANGELOG.md
@@ -42,6 +42,13 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 * [#593](https://github.com/provenance-io/cosmos-sdk/pull/593) Provenance: Use a new IMsgServiceRouter interface instead of a concrete one.
 * [#594](https://github.com/provenance-io/cosmos-sdk/pull/594) Provenance: Create the InputOutputCoinsProv function that's similar to InputOutputCoins but allows many inputs with a single output (as well as 1-1 and 1-many still).
+* [#595](https://github.com/provenance-io/cosmos-sdk/pull/595) Provenance: Add ability to inject a locked coins function into the bank module.
+
+### Improvements
+
+* [#595](https://github.com/provenance-io/cosmos-sdk/pull/595) Provenance: Create SimulateFromSeedProv which is the same as SimulateFromSeed but also returns the simulation's end block time which is needed for some sims tests.
+
+---
 
 ## [v0.50.4](https://github.com/cosmos/cosmos-sdk/releases/tag/v0.50.4) - 2023-02-19